Try using onPointerDown for Moveable.
