+++ assets/src/edit-story/elements/image.js
@@ -16,7 +16,7 @@ const Element = styled.img`
 	${ ElementWithRotation }
 `;
 
-function Image( { src, width, height, x, y, rotationAngle, forwardedRef } ) {
+function Image( { src, width, height, x, y, rotationAngle, forwardedRef, onPointerDown } ) {
 	const props = {
 		width,
 		height,
@@ -27,7 +27,7 @@ function Image( { src, width, height, x, y, rotationAngle, forwardedRef } ) {
 		ref: forwardedRef,
 	};
 	return (
-		<Element ref={ forwardedRef } { ...props } />
+		<Element { ...props } onPointerDown={ onPointerDown } />
 	);
 }
